fixed accessibility focused on chips
