--- browse/src/main/kotlin/com/alfresco/content/browse/tasks/ChipFilterType.kt
@@ -1,12 +1,0 @@
-package com.alfresco.content.browse.tasks
-
-/**
- * This is chip component of various types
- */
-enum class ChipFilterType(val component: String) {
-
-    TEXT("text"),
-    DATE_RANGE("date-range"),
-    RADIO("radio"),
-    None("none")
-}
