+++ common/src/main/res/values-es/strings.xml
@@ -3,16 +3,4 @@
     <string name="permission_rationale_title">Se necesitan permisos</string>
     <string name="permission_rationale_positive">Aceptar</string>
     <string name="permission_rationale_negative">Cancelar</string>
-
-    <string name="search_hint">Buscar</string>
-    <string name="advance_search_title">Búsqueda avanzada</string>
-    <string name="hint_range_from">De</string>
-    <string name="hint_range_to">A</string>
-    <string name="hint_range_from_date">Desde la fecha</string>
-    <string name="hint_range_to_date">Hasta la fecha</string>
-    <string name="component_number_range_invalid_input">El valor De debe ser menor que el valor A</string>
-    <string name="component_action_reset">Restablecer</string>
-    <string name="component_action_apply">Aplicar</string>
-    <string name="component_action_cancel">Cancelar</string>
-    <string name="component_number_range_empty">Valor requerido</string>
 </resources>
