+++ common/src/main/res/values-nl/strings.xml
@@ -3,16 +3,4 @@
     <string name="permission_rationale_title">Rechten nodig</string>
     <string name="permission_rationale_positive">OK</string>
     <string name="permission_rationale_negative">Annuleren</string>
-
-    <string name="search_hint">Zoeken</string>
-    <string name="advance_search_title">Geavanceerd zoeken</string>
-    <string name="hint_range_from">Van</string>
-    <string name="hint_range_to">Tot</string>
-    <string name="hint_range_from_date">Van-datum</string>
-    <string name="hint_range_to_date">Tot-datum</string>
-    <string name="component_number_range_invalid_input">Van-waarde moet kleiner zijn dan Tot-waarde</string>
-    <string name="component_action_reset">Opnieuw instellen</string>
-    <string name="component_action_apply">Toepassen</string>
-    <string name="component_action_cancel">Annuleren</string>
-    <string name="component_number_range_empty">Vereiste waarde</string>
 </resources>
