+++ component/src/main/java/com/alfresco/content/component/ComponentMetaData.kt
@@ -5,5 +5,6 @@ package com.alfresco.content.component
  */
 data class ComponentMetaData(
     val name: String? = "",
-    val query: String? = ""
+    val query: String? = "",
+    val queryMap: Map<String, String>? = mapOf()
 )
