+++ component/src/main/res/layout/view_list_sort_chips.xml
@@ -1,14 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<merge xmlns:android="http://schemas.android.com/apk/res/android">
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <FrameLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content">
 
-        <com.alfresco.content.FilterChip
+        <com.alfresco.content.component.FilterChip
             android:id="@+id/chip"
             style="@style/Widget.Alfresco.Chip.Choice"
             android:layout_width="wrap_content"
+            app:singleSelection = "true"
             android:layout_height="wrap_content"
             android:layout_marginStart="12dp" />
