+++ component/src/main/res/values-es/strings.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="search_hint">Buscar</string>
+    <string name="advance_search_title">Búsqueda avanzada</string>
+    <string name="hint_range_from">De</string>
+    <string name="hint_range_to">A</string>
+    <string name="hint_range_from_date">Desde la fecha</string>
+    <string name="hint_range_to_date">Hasta la fecha</string>
+    <string name="component_number_range_invalid_input">El valor De debe ser menor que el valor A</string>
+    <string name="component_action_reset">Restablecer</string>
+    <string name="component_action_apply">Aplicar</string>
+    <string name="component_action_cancel">Cancelar</string>
+    <string name="component_number_range_empty">Valor requerido</string>
+    <string name="size_end_kb">%1$s (KB)</string>
+</resources>
