+++ component/src/main/res/values-fr/strings.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="search_hint">Rechercher</string>
+    <string name="advance_search_title">Recherche avancée</string>
+    <string name="hint_range_from">De</string>
+    <string name="hint_range_to">A</string>
+    <string name="hint_range_from_date">Du (date)</string>
+    <string name="hint_range_to_date">Au (date)</string>
+    <string name="component_number_range_invalid_input">La valeur \'De\' doit être inférieure à la valeur \'A\'</string>
+    <string name="component_action_reset">Réinitialiser</string>
+    <string name="component_action_apply">Appliquer</string>
+    <string name="component_action_cancel">Annuler</string>
+    <string name="component_number_range_empty">Valeur requise</string>
+    <string name="size_end_kb">%1$s (Ko)</string>
+</resources>
