+++ component/src/main/res/values-it/strings.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="search_hint">Ricerca</string>
+    <string name="advance_search_title">Ricerca avanzata</string>
+    <string name="hint_range_from">Da</string>
+    <string name="hint_range_to">A</string>
+    <string name="hint_range_from_date">Data di inizio</string>
+    <string name="hint_range_to_date">Data di fine</string>
+    <string name="component_number_range_invalid_input">Il valore Da deve essere inferiore al valore A</string>
+    <string name="component_action_reset">Reimposta</string>
+    <string name="component_action_apply">Applica</string>
+    <string name="component_action_cancel">Annulla</string>
+    <string name="component_number_range_empty">Valore obbligatorio</string>
+    <string name="size_end_kb">%1$s (KB)</string>
+</resources>
