+++ component/src/main/res/values-night/colors.xml
