+++ component/src/main/res/values/colors.xml
