--- search/src/main/kotlin/com/alfresco/content/search/components/ComponentMetaData.kt
@@ -1,9 +1,0 @@
-package com.alfresco.content.search.components
-
-/**
- * Component Meta data for the return result
- */
-data class ComponentMetaData(
-    val name: String? = "",
-    val query: String? = ""
-)
