--- search/src/main/res/layout/sheet_component_create.xml
@@ -1,184 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:paddingTop="16dp"
-    android:paddingBottom="16dp"
-    android:layout_height="match_parent">
-
-
-    <LinearLayout
-        android:id="@+id/parent_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/component_default_margin"
-        android:layout_marginEnd="@dimen/component_default_margin"
-        android:orientation="vertical"
-        app:behavior_hideable="false">
-
-        <RelativeLayout
-            android:id="@+id/topView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <TextView
-                android:id="@+id/title"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_alignTop="@id/cancel_button"
-                android:layout_alignBottom="@id/cancel_button"
-                android:layout_toStartOf="@id/cancel_button"
-                android:ellipsize="end"
-                android:gravity="center_vertical"
-                android:singleLine="true"
-                android:textAppearance="?attr/textAppearanceSubtitle1"
-                tools:text="Title" />
-
-            <ImageView
-                android:id="@+id/cancel_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentEnd="true"
-                android:contentDescription="@string/component_action_cancel"
-                android:src="@drawable/ic_cancel_component" />
-        </RelativeLayout>
-
-        <View
-            android:id="@+id/separator"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:layout_marginTop="8dp"
-            android:background="@color/color_view_line" />
-
-
-        <FrameLayout
-            android:id="@+id/frameText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/textComponent"
-                layout="@layout/view_text_component" />
-        </FrameLayout>
-
-        <FrameLayout
-            android:id="@+id/frameRadio"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/radioListComponent"
-                layout="@layout/view_radio_list_component" />
-        </FrameLayout>
-
-        <FrameLayout
-            android:id="@+id/frameCheckList"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/checkListComponent"
-                layout="@layout/view_check_list_component" />
-        </FrameLayout>
-
-        <FrameLayout
-            android:id="@+id/frameNumberRange"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/numberRangeComponent"
-                layout="@layout/view_number_range_component" />
-        </FrameLayout>
-
-        <FrameLayout
-            android:id="@+id/frameSlider"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/sliderComponent"
-                layout="@layout/view_slider_component" />
-        </FrameLayout>
-
-        <FrameLayout
-            android:id="@+id/frameDateRange"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/dateRangeComponent"
-                layout="@layout/view_date_range_component" />
-        </FrameLayout>
-
-        <FrameLayout
-            android:id="@+id/frameFacet"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp">
-
-            <include
-                android:id="@+id/facetCheckListComponent"
-                layout="@layout/view_facet_check_list_component" />
-        </FrameLayout>
-
-
-    </LinearLayout>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_marginTop="16dp"
-        android:background="@color/color_view_line" />
-
-    <LinearLayout
-        android:id="@+id/bottom_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/component_default_margin"
-        android:layout_marginEnd="@dimen/component_default_margin"
-        android:orientation="vertical">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/component_default_margin"
-            android:layout_marginTop="16dp"
-            android:layout_marginEnd="@dimen/component_default_margin">
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/apply_button"
-                style="@style/Widget.Alfresco.Component.Primary"
-                android:layout_width="0dp"
-                android:layout_height="48dp"
-                android:layout_marginEnd="@dimen/component_button_margin"
-                android:text="@string/component_action_apply"
-                app:layout_constraintEnd_toStartOf="@id/reset_button"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/reset_button"
-                style="@style/Widget.Alfresco.Component.Secondary"
-                android:layout_width="0dp"
-                android:layout_height="48dp"
-                android:layout_marginStart="@dimen/component_button_margin"
-                android:backgroundTint="@color/color_component_secondary"
-                android:text="@string/component_action_reset"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@id/apply_button"
-                app:layout_constraintTop_toTopOf="parent" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </LinearLayout>
-
-</LinearLayout>
