+++ search/src/main/res/values-de/strings.xml
@@ -63,7 +63,6 @@
     <string name="category.default.search">Standard</string>
     <string name="category.folder">Ordner</string>
     <string name="type.facet.queries">Facet-Abfragen nach Typ</string>
-    <string name="size_end_kb">%1$s (KB)</string>
 
     <string name="search_move_warning">Sie können keine Dateien in Ihren eigenen Ordner und Unterordner verschieben.</string>
 </resources>
