+++ search/src/main/res/values-es/strings.xml
@@ -62,7 +62,6 @@
     <string name="category.default.search">Predeterminado</string>
     <string name="category.folder">Carpeta</string>
     <string name="type.facet.queries">Consultas con filtros por tipo</string>
-    <string name="size_end_kb">%1$s (KB)</string>
 
     <string name="search_move_warning">No puede mover ficheros a su propia carpeta y subcarpetas.</string>
