+++ search/src/main/res/values-fr/strings.xml
@@ -62,7 +62,6 @@
     <string name="category.default.search">Défaut</string>
     <string name="category.folder">Dossier</string>
     <string name="type.facet.queries">Recherches avec filtre par type</string>
-    <string name="size_end_kb">%1$s (Ko)</string>
 
     <string name="search_move_warning">Vous ne pouvez pas déplacer des fichiers vers votre propre dossier et sous-dossiers.</string>
