+++ search/src/main/res/values-it/strings.xml
@@ -62,7 +62,6 @@
     <string name="category.default.search">Predefinito</string>
     <string name="category.folder">Cartella</string>
     <string name="type.facet.queries">Query con filtro per tipo</string>
-    <string name="size_end_kb">%1$s (KB)</string>
 
     <string name="search_move_warning">Impossibile spostare i file nella cartella personale e nelle relative sottocartelle.</string>
