+++ search/src/main/res/values/strings.xml
@@ -62,6 +62,5 @@
     <string name="category.default.search">Default</string>
     <string name="category.folder">Folder</string>
     <string name="type.facet.queries">Type facet queries</string>
-    <string name="size_end_kb">%1$s (KB)</string>
     <string name="search_move_warning">you cannot move to own and subfolders.</string>
 </resources>
